refactor: product폴더 이하 모든 파일을 domain/product 폴더로 이동
