+++ src/main/java/com/example/WonkaoTalk/domain/product/entity/CartItem.java
@@ -1,9 +1,7 @@
-package com.example.WonkaoTalk.product.entity;
+package com.example.WonkaoTalk.domain.product.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -14,34 +12,28 @@ import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "product_variant")
-public class ProductVariant {
+@Table(name = "cart_item")
+public class CartItem {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "variant_id")
-  private Long variantId;
+  @Column(name = "cart_item_id")
+  private Long cartItemId;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "product_id", nullable = false)
-  private Product product;
+  @JoinColumn(name = "cart_id", nullable = false)
+  private Cart cart;
 
-  @Column(name = "stock", nullable = false)
-  private Integer stock;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "variant_id", nullable = false)
+  private ProductVariant productVariant;
 
-  @Column(name = "variant_name")
-  private String variantName;
-
-  @Enumerated(EnumType.STRING)
-  @Column(name = "status", nullable = false)
-  private SaleStatus status;
+  @Column(name = "quantity", nullable = false)
+  private Integer quantity;
 
   @Column(name = "created_at", nullable = false)
   private LocalDateTime createdAt;
 
   @Column(name = "updated_at", nullable = false)
   private LocalDateTime updatedAt;
-
-  @Column(name = "deleted_at")
-  private LocalDateTime deletedAt;
 }
