+++ src/main/java/com/example/WonkaoTalk/domain/product/entity/Category.java
@@ -1,4 +1,4 @@
-package com.example.WonkaoTalk.product.entity;
+package com.example.WonkaoTalk.domain.product.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -11,18 +11,21 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
 @Entity
-@Table(name = "product_option")
-public class ProductOption {
+@Table(name = "category")
+public class Category {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "product_option_id")
-  private Long productOptionId;
+  @Column(name = "category_id")
+  private Long categoryId;
 
-  @Column(name = "option_name", nullable = false)
-  private String optionName;
+  @Column(name = "category_name", nullable = false)
+  private String categoryName;
+
+  @Column(name = "depth", nullable = false)
+  private Integer depth;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "product_option_group_id", nullable = false)
-  private ProductOptionGroup productOptionGroup;
+  @JoinColumn(name = "parent_category_id")
+  private Category parentCategory;
 }
