+++ src/main/java/com/example/WonkaoTalk/domain/product/entity/ProductVariant.java
@@ -1,7 +1,9 @@
-package com.example.WonkaoTalk.product.entity;
+package com.example.WonkaoTalk.domain.product.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -12,28 +14,34 @@ import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "cart_item")
-public class CartItem {
+@Table(name = "product_variant")
+public class ProductVariant {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "cart_item_id")
-  private Long cartItemId;
+  @Column(name = "variant_id")
+  private Long variantId;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "cart_id", nullable = false)
-  private Cart cart;
+  @JoinColumn(name = "product_id", nullable = false)
+  private Product product;
 
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "variant_id", nullable = false)
-  private ProductVariant productVariant;
+  @Column(name = "stock", nullable = false)
+  private Integer stock;
 
-  @Column(name = "quantity", nullable = false)
-  private Integer quantity;
+  @Column(name = "variant_name")
+  private String variantName;
+
+  @Enumerated(EnumType.STRING)
+  @Column(name = "status", nullable = false)
+  private SaleStatus status;
 
   @Column(name = "created_at", nullable = false)
   private LocalDateTime createdAt;
 
   @Column(name = "updated_at", nullable = false)
   private LocalDateTime updatedAt;
+
+  @Column(name = "deleted_at")
+  private LocalDateTime deletedAt;
 }
