+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/CartItemRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.CartItem;
+import com.example.WonkaoTalk.domain.product.entity.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CartItemRepository extends JpaRepository<CartItem, Long> {
+
 }
