+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/CartRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.Cart;
+import com.example.WonkaoTalk.domain.product.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CartRepository extends JpaRepository<Cart, Long> {
+
 }
