+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/ProductDetailRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.ProductDetail;
+import com.example.WonkaoTalk.domain.product.entity.ProductDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductDetailRepository extends JpaRepository<ProductDetail, Long> {
+
 }
