+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/ProductImageRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.ProductImage;
+import com.example.WonkaoTalk.domain.product.entity.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+
 }
