+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/ProductOptionGroupRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.ProductOptionGroup;
+import com.example.WonkaoTalk.domain.product.entity.ProductOptionGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductOptionGroupRepository extends JpaRepository<ProductOptionGroup, Long> {
+
 }
