+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/ProductOptionRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.ProductOption;
+import com.example.WonkaoTalk.domain.product.entity.ProductOption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductOptionRepository extends JpaRepository<ProductOption, Long> {
+
 }
