+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/ProductRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.Product;
+import com.example.WonkaoTalk.domain.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
+
 }
