+++ src/main/java/com/example/WonkaoTalk/domain/product/repo/ProductVariantRepository.java
@@ -1,7 +1,8 @@
-package com.example.WonkaoTalk.product.repo;
+package com.example.WonkaoTalk.domain.product.repo;
 
-import com.example.WonkaoTalk.product.entity.ProductVariant;
+import com.example.WonkaoTalk.domain.product.entity.ProductVariant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductVariantRepository extends JpaRepository<ProductVariant, Long> {
+
 }
